Improve recommendation speed
We don't need to compute all permutations of the people available.We just need all possible pairing settings in any order.
To achieve that we recursively find pairings swapping only two persons at a time.
Then we proceed with the same previous logic of trying all the possibilities that are valid given a score.