Ensure that context is preserved based on yesterday
- It was possible before for track recommendation to behave strangely if
both members of a pair have worked on a track one time.  It could assign
the pair to work on the track even if they did not have context from the
previous day.  Now, this should not happen, since we set to max score if
one or the other member of the pair does not have context from the last
day.