Replace list item delete button with loading spinner
Before this change, when loading, a list item would
get a new spinner icon added. This slightly threw
off the layout while the spinner was present, and
it was more noticeable with the new emoji buttons.

This change makes it so the delete button gets replaced
briefly by the spinner.