Add initial recommendations code
This code is copied verbatim from the original Pairist codebase (plus or
minus a few formatting tweaks). I don't really understand how it works,
so I want to bring it over as-is for now.

It's not yet used in this app (the 'Recommend' button is not working),
but these tests are passing in isolation.

The next steps will be to adapt our slightly changed data models so they
can be plugged into this algorithm.