Implement pairing recommendations
In a previous commit, we brought the old recommendation algorithm logic
from Pairist v1, mostly unchanged. In this commit, we plug it into the
app. The tricky part here is that the shape of our data changed between
v1 and v2 (mainly due to differences between Firebase's Realtime
Database and Firestore). However, all the necessary info was still
present; it was just in a slightly different place.

Here, we introduce some adapter functions that take the new data
structures and convert them to the old, so they can be passed directly
into the recommendation-generation code. (I still don't really
understand it, so leaving it alone as much as possible!)

And finally, with this commit I'm adding a Notification/useNotification
pattern that can be used for temporary "toast"-style notifications – for
example, error messages that should be shown briefly to the user before
disappearing.