Allow reordering of lists
Lists can now be reordered with up/down arrow buttons that appear next
to the list titles.

Drag & drop would have been much trickier to implement given the way
list orders are stored in the database: we don't store an index for each
one, but rather the Date.now() at the time of creation. This allows us
to order by this field easily, but it means that we're not guaranteed
that these order numbers are always one-after-the-other.

Instead of arbitrary drag/drop, with these arrows we allow a simple
"swapping" of order in either direction (up or down). The moved list
just swaps its order number with the one above/below it.