Allow people to be locked in unassigned state
This adds a lock/unlock button on people when they are unassigned. When
locked, the recommendation algorithm will consider them 'out', so they
won't be placed in lanes.

This also adds a (currently not functional) delete button to people. A
follow-up commit will make this button remove a person from a team.