New function for API access
This is the first step towards enabling API access for teams. The goal
is to provide a GET endpoint that takes a team ID and (as a header) an
API key that is specific to that team. It returns the current pairs for
the team.

This is meant for clients that integrate with Pairist to consume (e.g.
Slack bots or CLI tools). It is intentionally limited in what it can do:
it is a read-only view into the currently assigned lanes.