Reimplement API endpoint for current pairs
Changing the approach to this. The original idea was to have API keys
stored for each team that could be used as headers in HTTP requests to
"authenticate". However, there were too many security questions around
this: although the data returned wasn't really sensitive, these API keys
were secrets, and there isn't a simple way to store secrets in the
Firebase DB. After thinking harder, there's a much simpler way to
achieve a similar thing...

Now, the API endpoint (".../current/<teamId>") is authenticated based on
your Firebase credentials – you'll need to be logged in as a real user
and pass an Authorization header with a valid Firebase ID token. For
teams that want to give programmatic access to certain users, the way to
do this is just to invite them to the team! This functionality already
exists and works fine.