Show helpful error when on invalid team page
Previously, if you visited a page like `/teams/bad`, where that team
eiteher didn't exist or you weren't a member, you'd just be on a blank
team forever (since the data would never load in).

Now, we redirect back to the home page in this case and show a brief
notification with a helpful message.