Switch to custom claim for email validation
Previously, we were using Firebase's standard emailVerified field to
keep track of whether or not this user had a valid email. However, this
is not sufficient because the email validation flow can be triggered by
any authenticated user – so they could verify themselves.

Switching to a custom claim means that only the admin can update this
field.