Add emoji reactions on list items
List items now have a button on them that pulls up an emoji picker.

For now, at least, we only support one of each emoji on each list item.
We store a count for each one (that's either 1 or 0), so we could add
support for more later if we want to.

Clicking on an emoji removes it.