Prevent flash of unverified screen on page load
Flipping the logic around here so that the result is pretty much
equivalent, but to eliminate a brief flash of the "Please verify your
email" screen when the page loads (even if you're verified). This is
because `auth.currentUser` is briefly undefined while the page loads –
and this was being considered "unverified" in the old logic.

Now, the logic is more accurate and we don't have that flash.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>