Add API endpoint for current lists
Also, refactor the current pairs endpoint to share a lot of code with
the new endpoint.

This new endpoint looks like `GET .../api/lists/:teamId` and (like the
other one) requires a valid ID token of a team member. It returns a JSON
representation of the current lists and list items. See the updated
docs/api.md page for details.