Upgrade react-markdown to fix link render bug
There was a bug where, if a list item contained text like "[RM]", the
markdown renderer would render it as a link with no href. Upgrading the
react-markdown dependency seems to fix this problem. Links that look
like "[text](url)" are still rendered as links, but only "[text]" will
be rendered verbatim, with the brackets shown.