Use early returns to handle emoji and color menu scrim clicks
Previously the color selector was written to use event handlers in
order to prevent the modal from closing on selection of a color. However
this created a bug that prevented the emoji menu from being able to open
after the color selector had been used. This commit switches to using an
early return based on active element class names when handling mouse
clicks to prevent this issue from occuring.