Unlock people when moving into a lane
We (relatively) recently added the feature where you can lock a person
by dragging them into the "locked" section of the "People" box on the
right. Doing this changes the person's `isLocked` property in the
database.

But we then allow the person to be dragged manually into a lane. When
this happens, we were not updating the `isLocked` property on the
person. This could be seen when sweeping a formerly-locked person out of
their lane: they would return to the "People" box in the locked state.

This also affected the recommendation logic:
https://github.com/pivotal-cf/pairist/blob/8f4ca005ecfec50f26377cc6f3961f993256b792/src/lib/adapter.ts#L54

That line means any people with `isLocked: true` would be considered
"out" (unavailable for pairing) for the purposes of the recommendation
engine. That throws things off, which is very unexpected if you've just
dragged the person into a lane and expect them to be paired.

We're fixing this by just updating the `isLocked` property to `true`
whenever we move a person into a lane. The logic here is a little tricky
because we don't want to just to `isLocked: !!laneId` - that would set
`isLocked` to true whenever we move a person out of a lane.